refactor(create-mc-app): to run tasks using listr output
