--- packages/create-mc-app/src/logger.js
@@ -1,13 +1,0 @@
-const { Signale } = require('signale');
-
-module.exports = new Signale({
-  config: { displayBadge: false, displayLabel: false },
-  scope: 'create-mc-app',
-  types: {
-    log: { color: '' },
-    success: { color: 'green' },
-    error: { color: 'red' },
-    info: { color: 'blue' },
-    note: { color: 'magenta' },
-  },
-});
