+++ packages/create-mc-app/src/utils.js
@@ -1,14 +1,10 @@
-const execSync = require('child_process').execSync;
+const execa = require('execa');
 
 const isSemVer = version => /^(v?)([0-9].[0-9].[0-9])+/.test(version);
 
 const shouldUseYarn = () => {
-  try {
-    execSync('yarnpkg --version', { stdio: 'ignore' });
-    return true;
-  } catch (e) {
-    return false;
-  }
+  const result = execa.sync('yarnpkg', ['--version'], { stdio: 'ignore' });
+  return !result.failed;
 };
 
 const slugify = name => name.toLowerCase().replace(/_/gi, '-');
